refactor: 파일위치 및 파일구조 수정
