+++ OpenMarket/OpenMarket/Data/OpenMarketRequest.swift
@@ -60,23 +60,3 @@ extension OpenMarketRequest {
     }
 }
 
-extension UIImage {
-    func resize(width: CGFloat) -> UIImage {
-        let scale = width / self.size.width
-        let newHeight = self.size.height * scale
-        
-        let size = CGSize(width: width, height: newHeight)
-        let render = UIGraphicsImageRenderer(size: size)
-        var renderImage = render.image { context in
-            self.draw(in: CGRect(origin: .zero, size: size))
-        }
-        
-        let imgData = NSData(data: renderImage.pngData()!)
-        let imageSize = Double(imgData.count) / 1000
-        
-        if imageSize > 300 {
-            renderImage = resize(width: width - 5)
-        }
-        return renderImage
-    }
-}
