+++ OpenMarket/OpenMarket/Extension/UIImage+extension.swift
@@ -1,0 +1,29 @@
+//
+//  UIImage.swift
+//  OpenMarket
+//
+//  Created by NAMU on 2022/08/05.
+//
+
+import UIKit.UIImage
+
+extension UIImage {
+    func resize(width: CGFloat) -> UIImage {
+        let scale = width / self.size.width
+        let newHeight = self.size.height * scale
+        
+        let size = CGSize(width: width, height: newHeight)
+        let render = UIGraphicsImageRenderer(size: size)
+        var renderImage = render.image { context in
+            self.draw(in: CGRect(origin: .zero, size: size))
+        }
+        
+        let imgData = NSData(data: renderImage.pngData()!)
+        let imageSize = Double(imgData.count) / 1000
+        
+        if imageSize > 300 {
+            renderImage = resize(width: width - 5)
+        }
+        return renderImage
+    }
+}
