+++ OpenMarket/OpenMarket/OpenMarketRequest/URLHost.swift
@@ -1,0 +1,18 @@
+//
+//  URLHost.swift
+//  OpenMarket
+//
+//  Created by groot, bard on 2022/08/05.
+//
+
+enum URLHost {
+    case openMarket
+    
+    var url: String {
+        switch self {
+        case .openMarket:
+            return "https://market-training.yagom-academy.kr"
+        }
+    }
+}
+
